requirements.txt
kafka-python
pyjnius
